feat(MineSweeper) : CLEAN - 패키지
