+++ src/main/kotlin/model/BoardSize.kt
@@ -1,3 +1,5 @@
+package model
+
 class BoardSize(private val row: LengthOfSide, private val col: LengthOfSide) {
 
     fun get(): Int = row.length * col.length
