+++ src/main/kotlin/model/MineType.kt
@@ -1,3 +1,5 @@
+package model
+
 enum class MineType(val symbol: Char) {
     NONE('C'),
     MINE('*')
