+++ src/main/kotlin/view/ResultView.kt
@@ -1,3 +1,8 @@
+package view
+
+import model.Board
+import model.LengthOfSide
+
 object ResultView {
 
     fun printBoard(board: Board, row: LengthOfSide) {
