+++ src/test/kotlin/model/BoardTest.kt
@@ -1,3 +1,5 @@
+package model
+
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
