+++ src/test/kotlin/model/LengthOfSideTest.kt
@@ -1,16 +1,16 @@
+package model
+
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 
-class NumberOfMineTest {
+class LengthOfSideTest {
 
     @ParameterizedTest
-    @ValueSource(ints = [0, 101])
-    fun `지뢰 개수가 유효하지 않은 경우`(input: Int) {
-        val boardSize = BoardSize(LengthOfSide(10), LengthOfSide(10))
-
+    @ValueSource(ints = [9, 31])
+    fun `높이 혹은 너비 값이 허용되는 범위에 없는 경우`(input: Int) {
         assertThatIllegalArgumentException().isThrownBy {
-            NumberOfMine(input, boardSize)
+            LengthOfSide(input)
         }
     }
 }
