+++ src/test/kotlin/model/NumberOfMineTest.kt
@@ -1,14 +1,18 @@
+package model
+
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 
-class LengthOfSideTest {
+class NumberOfMineTest {
 
     @ParameterizedTest
-    @ValueSource(ints = [9, 31])
-    fun `높이 혹은 너비 값이 허용되는 범위에 없는 경우`(input: Int) {
+    @ValueSource(ints = [0, 101])
+    fun `지뢰 개수가 유효하지 않은 경우`(input: Int) {
+        val boardSize = BoardSize(LengthOfSide(10), LengthOfSide(10))
+
         assertThatIllegalArgumentException().isThrownBy {
-            LengthOfSide(input)
+            NumberOfMine(input, boardSize)
         }
     }
 }
